Added config file for diode power ranges.

+++ src/main/java/com/johnwigg/laserdiodedriverui/RangeSettingListener.java
@@ -1,5 +1,5 @@
 package com.johnwigg.laserdiodedriverui;
 
 public interface RangeSettingListener {
-	void onRangeSetting(int index, boolean is_max, float value);
+	void onRangeSetting(int index, boolean is_max, int value);
 }
